+++ client/galaxy/scripts/mvc/collection/ListCollectionCreatorModal.vue
@@ -1,0 +1,52 @@
+<script>
+import CollectionCreatorModalMixin from "./mixins/CollectionCreatorModalMixin";
+import ListCollectionCreator from "./ListCollectionCreator";
+import _l from "utils/localization";
+import Vue from "vue";
+export default {
+    mixins: [
+        CollectionCreatorModalMixin,
+    ],
+    methods: {
+        listCollectionCreatorModal: function (elements, options) {
+            options = options || {};
+            options.title = _l("Create a collection from a list of datasets");
+            const { deferred, creatorOptions, showEl } = this.collectionCreatorModalSetup(options); // eslint-disable-line no-unused-vars
+
+            var vm = document.createElement("div");
+            showEl(vm);
+            var listCollectionCreatorClass = Vue.extend(ListCollectionCreator);
+            var listinstance = new listCollectionCreatorClass({
+                propsData: {
+                    creationFn: options.creationFn,
+                    oncancel: options.oncancel,
+                    oncreate: options.oncreate,
+                    //TODO : autoscollDist, highlightClr
+                },
+            });
+            listinstance.$mount(vm);
+            return deferred;
+        },
+        /** Use a modal to create a list collection, then add it to the given history contents.
+         *  @returns {Deferred} resolved when the collection is added to the history.
+         */
+        createListCollection: function (contents, defaultHideSourceItems) {
+            const elements = contents.toJSON();
+            const copyElements = !defaultHideSourceItems;
+            const promise = this.listCollectionCreatorModal(elements, {
+                defaultHideSourceItems: defaultHideSourceItems,
+                creationFn: function (elements, name, hideSourceItems) {
+                    elements = elements.map((element) => ({
+                        id: element.id,
+                        name: element.name,
+                        //TODO: this allows for list:list even if the filter above does not - reconcile
+                        src: element.history_content_type === "dataset" ? "hda" : "hdca",
+                    }));
+                    return contents.createHDCA(elements, "list", name, hideSourceItems, copyElements);
+                },
+            });
+            return promise;
+        },
+    },
+};
+</script>
